Move AvatarUploader to organism since it has a very heavy tool that is cropTool
